Create 1st Session (15.12.2018)
During the first one hour meeting, we introduced aims and goals of the group, what level of experience everyone has with python and what is hoped to be achieved. We discussed potential projects to work on over the coming weeks. 
The most helpful links to begin learning Python from the basics include:
online courses
https://www.codecademy.com/learn/learn-python
https://www.sololearn.com/Course/Python/
https://www.udemy.com/the-python-bible/learn/v4/content
https://www.python-course.eu/
http://www.learnpython.org/
https://www.datacamp.com/home
https://thehelloworldprogram.com/python/
https://www.dataquest.io/m/1/python-basics
https://app.finxter.com/learn/computer/science/
http://www.learnpython.org/en/Hello%2C_World%21
https://realpython.com/
https://groklearning.com/launch/#type=course
https://www.codewars.com/?language=python